+++ ios/Classes/internal/ZegoExpressEngineMethodHandler.m
@@ -347,6 +347,14 @@
     result(nil);
 }
 
+- (void)submitLog:(FlutterMethodCall *)call result:(FlutterResult)result {
+
+    [ZegoExpressEngine submitLog];
+
+    result(nil);
+}
+
+
 - (void)enableDebugAssistant:(FlutterMethodCall *)call result:(FlutterResult)result {
 
     BOOL enable = [ZegoUtils boolValue:call.arguments[@"enable"]];
@@ -702,6 +710,33 @@
     int channel = [ZegoUtils intValue:call.arguments[@"channel"]];
 
     [[ZegoExpressEngine sharedEngine] setVideoMirrorMode:(ZegoVideoMirrorMode)mode channel:(ZegoPublishChannel)channel];
+
+    result(nil);
+
+}
+
+- (void)setPublishDualStreamConfig:(FlutterMethodCall *)call result:(FlutterResult)result {
+
+    NSArray<NSDictionary *> *configListMap = call.arguments[@"configList"];
+    NSMutableArray<ZegoPublishDualStreamConfig *> *configList = [[NSMutableArray alloc] init];
+    
+    for (NSDictionary *configMap in configListMap) {
+        ZegoPublishDualStreamConfig *config = [[ZegoPublishDualStreamConfig alloc] init];
+        int streamType = [ZegoUtils intValue:configMap[@"streamType"]];
+        int width = [ZegoUtils intValue:configMap[@"encodeWidth"]];
+        int height = [ZegoUtils intValue:configMap[@"encodeHeight"]];
+        
+        config.streamType = (ZegoVideoStreamType) streamType;
+        config.fps = [ZegoUtils intValue:configMap[@"fps"]];
+        config.bitrate = [ZegoUtils intValue:configMap[@"bitrate"]];
+        config.encodeResolution = CGSizeMake(width, height);
+        
+        [configList addObject:config];
+    }
+    
+    int channel = [ZegoUtils intValue:call.arguments[@"channel"]];
+
+    [[ZegoExpressEngine sharedEngine] setPublishDualStreamConfig:configList.copy channel:(ZegoPublishChannel)channel];
 
     result(nil);
 
@@ -1108,6 +1143,9 @@
         hasChannel = YES;
         channel = [ZegoUtils intValue:call.arguments[@"channel"]];
     }
+    
+    /// 标识推流通道的视频源
+    [ZegoTextureRendererController.sharedInstance setVideoSourceChannel:@(channel) withSource:source];
 
     int ret = 0;
     if (!hasChannel && !hasInstanceID) {
@@ -1118,13 +1156,6 @@
         ret = [[ZegoExpressEngine sharedEngine] setVideoSource:(ZegoVideoSourceType)source instanceID:instanceID];
     } else {
         ret = [[ZegoExpressEngine sharedEngine] setVideoSource:(ZegoVideoSourceType)source instanceID:instanceID channel:(ZegoPublishChannel)channel];
-    }
-
-    /// 标识推流通道的视频源
-    if (source == ZegoVideoSourceTypeScreenCapture) {
-        _screenCaptureChannel = channel;
-    } else if (self.screenCaptureChannel == channel) {
-        _screenCaptureChannel = -1;
     }
 
     result(@(ret));